% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spilhaus.R
\name{spilhaus}
\alias{spilhaus}
\title{Project points to Spilhaus}
\usage{
spilhaus(x)
}
\arguments{
\item{x}{matrix of longlat}
}
\value{
matrix of projected coordinates
}
\description{
Input a matrix of longitude latitude coordinates and return Spilhaus coordinates.
}
\details{
Note that this transformation is not available via standard libraries, and was
implemented in R and Python by Ricardo T. Lemos.
}
\examples{
spilhaus(cbind(0, 0))
}
